pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Georgypython <nov_georgy@inbox.ru>"]
readme = "README.md"
packages = [{include = "hexlet_code"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
python-dotenv = "^1.0.0"
django-bootstrap5 = "^23.3"
flake8 = "^6.1.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
